login

+++ src/main/java/com/spring/travelmvc/CommentController.java
@@ -1,6 +1,10 @@
 package com.spring.travelmvc;
 
+import java.io.IOException;
 import java.util.ArrayList;
+
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,6 +15,7 @@ import com.spring.travelmvc.impl.CommentDao;
 import com.spring.travelmvc.impl.CommentDo;
 import com.spring.travelmvc.impl.TravelDao;
 import com.spring.travelmvc.impl.TravelDo;
+import com.spring.travelmvc.util.ScriptUtil;
 
 @Controller
 public class CommentController {
@@ -21,9 +26,14 @@ public class CommentController {
 	CommentDao commentDao;
 	
 	@RequestMapping(value="/insertProcComment.do")
-	public String insertProcComment(TravelDo tdo, CommentDo cdo, Model model) {
-		commentDao.insertComment(tdo, cdo);
-		
+	public String insertProcComment(TravelDo tdo, CommentDo cdo, Model model, HttpSession session, HttpServletResponse response) throws IOException {
+		String sessionId = (String)session.getAttribute("sessionId");
+		if (sessionId == null) {
+			ScriptUtil.alertAndBackPage(response,  "로그인 후 이용해 주세요.");
+		} else {
+			cdo.setWriter(sessionId);
+			commentDao.insertComment(tdo, cdo);
+		}		
 		TravelDo travel = travelDao.getTravel(tdo);
 		ArrayList<CommentDo> cList = commentDao.getCommentList(tdo);
 		model.addAttribute("travel", travel);
@@ -33,25 +43,53 @@ public class CommentController {
 	}
 	
 	@RequestMapping(value="/deleteComment.do")
-	public String deleteComment(TravelDo tdo, CommentDo cdo, Model model) {
-		commentDao.deleteComment(cdo);
+	public String deleteComment(TravelDo tdo, CommentDo cdo, Model model, HttpSession session, HttpServletResponse response) throws IOException {
 		
 		TravelDo travel = travelDao.getTravel(tdo);
 		ArrayList<CommentDo> cList = commentDao.getCommentList(tdo);
-		model.addAttribute("travel", travel);
-		model.addAttribute("cList", cList);
+		String commentId = null;
+		for (CommentDo comment : cList) {
+			if (comment.getCommentSeq() == cdo.getCommentSeq()) {
+				commentId = comment.getWriter();
+			}
+		}
+		String sessionId = (String)session.getAttribute("sessionId");
+		
+		if (sessionId == null) {
+			ScriptUtil.alertAndBackPage(response,  "로그인 후 이용해 주세요.");
+		} else if (!sessionId.equals(commentId)) {
+			ScriptUtil.alertAndBackPage(response,  "타인의 글은 삭제할 수 없습니다.");
+		} else {
+			commentDao.deleteComment(cdo);
+			cList = commentDao.getCommentList(tdo);
+			model.addAttribute("travel", travel);
+			model.addAttribute("cList", cList);
+		}
 		
 		return "getTravelView";
 	}
 	
 	@RequestMapping(value="/modifyComment.do")
-	public String modifyComment(TravelDo tdo, CommentDo cdo, Model model) {
+	public String modifyComment(TravelDo tdo, CommentDo cdo, Model model, HttpSession session, HttpServletResponse response) throws IOException {
 		TravelDo travel = travelDao.getTravel(tdo);
 		ArrayList<CommentDo> cList = commentDao.getCommentList(tdo);
 		model.addAttribute("travel", travel);
 		model.addAttribute("cList", cList);
 		model.addAttribute("cdo", cdo);
 		
+		String commentId = null;
+		for (CommentDo comment : cList) {
+			if (comment.getCommentSeq() == cdo.getCommentSeq()) {
+				commentId = comment.getWriter();
+			}
+		}
+		String sessionId = (String)session.getAttribute("sessionId");
+		
+		if (sessionId == null) {
+			ScriptUtil.alertAndBackPage(response,  "로그인 후 이용해 주세요.");
+		} else if (!sessionId.equals(commentId)) {
+			ScriptUtil.alertAndBackPage(response,  "타인의 글은 수정할 수 없습니다.");
+		}
 		return "getModifyCommentView";
 	}
 	
@@ -69,5 +107,5 @@ public class CommentController {
 		return "getTravelView";
 	}
 	
-	
+	// TODO: 작성자만 댓글 삭제할 수 있도록, 작성자만 게시글 수정/삭제할 수 있도록
 }
