+++ src/main/java/com/spring/travelmvc/TravelController.java
@@ -1,6 +1,10 @@
 package com.spring.travelmvc;
 
+import java.io.IOException;
 import java.util.ArrayList;
+
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -12,6 +16,7 @@ import com.spring.travelmvc.impl.CommentDao;
 import com.spring.travelmvc.impl.CommentDo;
 import com.spring.travelmvc.impl.TravelDao;
 import com.spring.travelmvc.impl.TravelDo;
+import com.spring.travelmvc.util.ScriptUtil;
 
 @Controller
 public class TravelController {
@@ -28,6 +33,10 @@ public class TravelController {
 	
 	@RequestMapping(value="/insertProcTravel.do")
 	public String insertProcTravel(TravelDo tdo) {
+		
+		System.out.println(tdo.getTitle());
+		System.out.println(tdo.getWriter());
+		System.out.println(tdo.getContent());
 		travelDao.insertTravel(tdo);
 		
 		return "redirect:getTravelList.do";
@@ -52,9 +61,16 @@ public class TravelController {
 	}
 	
 	@RequestMapping(value="/modifyTravel.do")
-	public String modifyTravel(TravelDo tdo, Model model) {
+	public String modifyTravel(TravelDo tdo, Model model, HttpSession session, HttpServletResponse response) throws IOException {
 		TravelDo travel = travelDao.getTravel(tdo);
 		model.addAttribute("travel", travel);
+		TravelDo dbTdo = travelDao.getTravel(tdo);
+		String sessionId = (String)session.getAttribute("sessionId");
+		if (sessionId == null) {
+			ScriptUtil.alertAndBackPage(response,  "로그인 후 이용해 주세요.");
+		} else if (!sessionId.equals(dbTdo.getWriter())) {
+			ScriptUtil.alertAndBackPage(response,  "타인의 글은 삭제할 수 없습니다.");
+		}
 		
 		return "getModifyView";
 	}
@@ -67,8 +83,16 @@ public class TravelController {
 	}
 	
 	@RequestMapping(value="/deleteTravel.do")
-	public String deleteBTravel(TravelDo tdo, Model model) {
-		travelDao.deleteTravel(tdo);
+	public String deleteBTravel(TravelDo tdo, Model model, HttpSession session, HttpServletResponse response) throws IOException {
+		TravelDo dbTdo = travelDao.getTravel(tdo);
+		String sessionId = (String)session.getAttribute("sessionId");
+		if (sessionId == null) {
+			ScriptUtil.alertAndBackPage(response,  "로그인 후 이용해 주세요.");
+		} else if (!sessionId.equals(dbTdo.getWriter())) {
+			ScriptUtil.alertAndBackPage(response,  "타인의 글은 삭제할 수 없습니다.");
+		} else {
+			travelDao.deleteTravel(tdo);
+		}
 		
 		return "redirect:getTravelList.do";
 	}
